requirements.txt
# Core dependencies
pandas>=2.2.0
numpy>=1.26.0

# Environment
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Machine Learning (Phase 2)
scikit-learn==1.3.2
scikit-surprise==1.1.3
scipy>=1.11.0

# CLI & Utilities
tabulate==0.9.0

# Web UI (Phase 2)
streamlit==1.29.0

# Jupyter
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.3
